Moved around ECS stage

+++ Jenkinsfile
@@ -27,8 +27,8 @@ pipeline {
         USERS_IMAGE = "$ECR_URI/ab-users-microservice:latest"
         BOOKINGS_IMAGE = "$ECR_URI/ab-bookings-microservice:latest"
 
-        PROJECT_ID = credentials('project-id')
-        ENV = credentials('env')
+        PROJECT_ID  = credentials('project-id')
+        ENV         = credentials('env')
         PUB_SSH_KEY = credentials('pub-ssh-key')
     }
 
@@ -78,84 +78,46 @@ EOF
             steps {
                 dir("terraform") {
                     sh "terraform apply -no-color -input=false plans/apply-${COMMIT_HASH}.tf"
-                    sh 'terraform refresh'
+                    sh 'terraform refresh -no-color'
                     sh 'terraform output | tr -d \'\\\"\\ \' > output.tf'
                 }
             }
         }
 
-        stage('Deploy ECS cluster') {
-            when {
-                expression {
-                    params.Apply
-                }
-            }
-
-            steps {
-                dir("ecs-${PROJECT_ID}") {
-                    script {
-                        withCredentials([
-                            string(
-                                credentialsId: "${ENV}/${PROJECT_ID}/default",
-                                variable: 'SECRETS'
-                            )
-                        ]) {
-                            def aws_secrets = readJSON text: SECRETS
-                            env.DB_USERNAME = aws_secrets.db_username
-                            env.DB_PASSWORD = aws_secrets.db_password
-                            env.JWT_SECRET  = aws_secrets.jwt_secret
-
-                            def tf_outputs = readProperties(file: '../terraform/output.tf')
-                            env.DOMAIN = tf_output.domain
-                            env.VPC_ID = tf_output.vpc_id
-                            env.DB_URL = tf_output.db_url
-                            env.ALB_ID = tf_output.alb_id
-
-                            sh "docker context use ecs-${PROJECT_ID}"
-                            sh "aws ecr get-login-password --region ${AWS_REGION} | docker login --username AWS --password-stdin ${AWS_ACCOUNT_ID}.dkr.ecr.${AWS_REGION}.amazonaws.com"
-                            sh "docker compose up --no-color"
-                        }
-                    }
-                }
-            }
-        }
-
-        stage('Destroy ECS cluster') {
-            when {
-                expression {
-                    params.Destroy
-                }
-            }
-
+        stage('ECS cluster') {
             steps {
                 script {
+                    withCredentials([
+                        string(
+                            credentialsId: "${ENV}/${PROJECT_ID}/default",
+                            variable: 'SECRETS'
+                        )
+                    ]) {
+                        def aws_secrets = readJSON text: SECRETS
+                        env.DB_USERNAME = aws_secrets.db_username
+                        env.DB_PASSWORD = aws_secrets.db_password
+                        env.JWT_SECRET  = aws_secrets.jwt_secret
+                    }
+
+                    def tf_outputs = readProperties(file: '../terraform/output.tf')
+                    env.DOMAIN = tf_output.domain
+                    env.VPC_ID = tf_output.vpc_id
+                    env.DB_URL = tf_output.db_url
+                    env.ALB_ID = tf_output.alb_id
+
                     dir("ecs-${PROJECT_ID}") {
-                        withCredentials([
-                            string(
-                                credentialsId: "${ENV}/${PROJECT_ID}/default",
-                                variable: 'SECRETS'
-                            )
-                        ]) {
-                            def aws_secrets = readJSON(text: SECRETS)
-                            env.DB_USERNAME = aws_secrets.db_username
-                            env.DB_PASSWORD = aws_secrets.db_password
-                            env.JWT_SECRET  = aws_secrets.jwt_secret
-
-                            def tf_outputs = readProperties(file: '../terraform/output.tf')
-                            env.DOMAIN = tf_outputs.domain
-                            env.VPC_ID = tf_outputs.vpc_id
-                            env.DB_URL = tf_outputs.db_url
-                            env.ALB_ID = tf_output.alb_id
-
-                            sh "docker context use ecs-${PROJECT_ID}"
-                            sh "aws ecr get-login-password --region ${AWS_REGION} | docker login --username AWS --password-stdin ${AWS_ACCOUNT_ID}.dkr.ecr.${AWS_REGION}.amazonaws.com"
+                        sh "docker context use ecs-${PROJECT_ID}"
+                        sh "aws ecr get-login-password --region ${AWS_REGION} | docker login --username AWS --password-stdin ${AWS_ACCOUNT_ID}.dkr.ecr.${AWS_REGION}.amazonaws.com"
+                        if(params.Apply) {
+                            sh "docker compose up --no-color"
+                        }
+                        if(params.Destroy) {
                             sh "docker compose down --no-color"
                         }
                     }
                 }
             }
         }
-
 
         stage('Terraform Plan Destroy') {
             when {
